app/requirements.txt
Flask==3.0.0
line-bot-sdk==2.4.2
gunicorn==21.2.0
